backend/requirements.txt
asgiref==3.11.1
cbor2==5.6.5
coverage==7.13.4
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-allauth==65.14.3
iniconfig==2.3.0
packaging==26.0
pillow==11.2.1
plugy==1.6.0
Pygments==2.19.2
PyJWT==2.11.0
pyotp==2.9.0
pytest==9.0.2
pytest-cov==7.0.0
pytest-django==4.12.0
python-decouple==3.8
qrcode==8.0
sqlparse==0.5.5
webauthn==2.3.0
